gre: Wait for an RCU grace period before freeing port.
We currently remove ports from the GRE hash table and then immediately
free the ports.  Since received packets could be using that port this
can lead to a crash (the port has already been detached from the
datapath so this can't happen for transmitted packets).  As a result
we need to wait for an RCU grace period to elapse before actually
freeing the port.

In an ideal world we would actually remove the port from the hash
table in a hypothetical gre_detach() function since this is one of
the purposes of detaching.  However, we also use the hash table to
look for collisions in the lookup criteria and don't want to allow
two identical ports to exist.  It doesn't matter though because we
aren't blocking on the freeing of resources.